pyproject.toml
[project]
name = "ai-reporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-frontmatter>=1.1.0",
]
